requirements.txt
argparse-actions==0.4.4
backports-abc==0.5
backports.ssl-match-hostname==3.7.0.1
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
click==8.1.3
cssselect==1.2.0
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.1.7
django-bootstrap5==22.2
django-braces==1.15.0
django-contrib-comments==2.2.0
django-currencies==0.10.1
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-extensions==3.2.1
django-filter==22.1
django-filebrowser-no-grappelli==4.0.1
django3-flatpages-tinymce==0.2.2
django-google-tag-manager==0.0.5
#django-google-analytics-app==5.0.2
django-oauth-plus==2.2.9
django-oauth-toolkit==2.2.0
django-oauth2-provider==0.2.6.1
django-simple-history==3.2.0
django-tempus-dominus==5.1.2.17
django-tinymce==3.5.0
django-toolbelt==0.0.1
django-uwsgi==0.2.2
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
drf_spectacular==0.25.1
filebrowser-safe==1.1.1
fontawesomefree==6.3.0
future==0.18.3
gunicorn==20.1.0
html5lib==1.1
httplib2==0.21.0
idna==3.4
mysql-connector-python==8.0.32
oauth2==1.9.0.post1
oauthlib==3.2.2
paypalrestsdk==1.13.1
pefile==2023.2.7
protobuf==4.22.0
Pillow==9.4.0
xhtml2pdf==0.2.9
pytz==2022.7.1
PyInstaller==5.8.0
PyPDF2==3.0.1
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
shortuuid==1.0.11
singledispatch==4.0.0
six==1.16.0
static3==0.7.0
tinycss==0.4
tornado==6.2
tzlocal==4.2
urllib3==1.26.14
wheel==0.38.4
paypalhttp==1.0.1
paypal-checkout-serversdk==1.0.1
django-cookiebanner==0.2.7
csscompressor==0.9.5
django-compressor==4.3.1
django-recaptcha==3.0.0
daphne==4.0.0
celery==5.2.7
redis==4.5.1
